+++ packages/vsce-api/CHANGELOG.md
@@ -2,9 +2,10 @@
 
 All notable changes to the IBM CICS for Zowe Explorer API will be documented in this file.
 
-## Recent Changes
+## `6.11.0`
 
 - Added new resource type - JVM server
+- Adds CICSBundle to list of supported resources
 
 ## `6.10.0`